Importer and db changes

+++ Tables/background_proficiency.sql
@@ -1,7 +1,7 @@
 CREATE TABLE background_proficiency (
     id INT PRIMARY KEY IDENTITY(1,1),
     background_id INT NOT NULL,
-    proficiency_id INT NOT NULL,
+    proficiency_id BIGINT NOT NULL,
     is_tool_proficiency BIT DEFAULT 0,
     created_at DATETIME DEFAULT GETDATE(),
     updated_at DATETIME DEFAULT GETDATE(),
