+++ Tables/class_proficiency_choices.sql
@@ -20,7 +20,7 @@ CREATE TABLE class_proficiency_choice_option
 (
     id INT PRIMARY KEY IDENTITY(1,1),
     choice_group_id INT NOT NULL,
-    proficiency_id INT NOT NULL,
+    proficiency_id BIGINT NOT NULL,
     option_type NVARCHAR(50) NOT NULL DEFAULT 'reference',
     created_at DATETIME DEFAULT GETDATE(),
     updated_at DATETIME DEFAULT GETDATE(),
