+++ Tables/trait_proficiency.sql
@@ -2,7 +2,7 @@ CREATE TABLE trait_proficiency
 (
     id INT PRIMARY KEY IDENTITY(1,1),
     trait_id INT NOT NULL,
-    proficiency_id INT NOT NULL,
+    proficiency_id BIGINT NOT NULL,
     created_at DATETIME DEFAULT GETDATE(),
     updated_at DATETIME DEFAULT GETDATE(),
     deleted BIT NULL,
